refactor: removed React imports & added SponsorPage styled component
