+++ src/pages/SponsorsPage/SponsorsPage.tsx
